@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 
 class CMScaffold extends StatelessWidget {
   final Widget body;
-  final String title;
+  final String? title;
   final List<Widget>? actions;
 
   const CMScaffold({
     required this.body,
-    required this.title,
+    this.title,
     this.actions,
     Key? key,
   }) : super(key: key);
@@ -16,7 +16,7 @@ class CMScaffold extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(title),
+        title: title != null ? Text(title!) : null,
         actions: actions,
       ),
       body: body,
