@@ -1,5 +1,21 @@
-extension DateTimeFormatting on DateTime {
+extension DateTimeExt on DateTime {
+  static DateTime? _mockedTime;
+
+  static DateTime now() => _mockedTime ?? DateTime.now();
+
+  static set mockTime(DateTime customTime) => _mockedTime = customTime;
+  static void resetMockTime() => _mockedTime = null;
+
   DateTime toStartOfDay() => DateTime(year, month, day);
+
+  DateTime nextWeekend() {
+    int daysUntilSaturday = 6 - weekday;
+    if (daysUntilSaturday < 1) {
+      daysUntilSaturday += 7;
+    }
+
+    return add(Duration(days: daysUntilSaturday)).toStartOfDay();
+  }
 
   String toDateString() {
     final String y =
