@@ -1,5 +1,6 @@
 import 'package:chore_manager_mobile/data/auth/auth_service.dart';
 import 'package:chore_manager_mobile/data/chore_manager_web/chores/chores_adapter.dart';
+import 'package:chore_manager_mobile/extensions/date_time_ext.dart';
 import 'package:chore_manager_mobile/modules/chores/chore.dart';
 import 'package:get/get.dart';
 
@@ -17,44 +18,75 @@ class ChoresController extends GetxController {
     await refreshChores();
   }
 
+  Chore chore(int id) => chores().firstWhere((chore) => chore.id == id);
+
   List<Chore> withDueDate() =>
-      chores().where((chore) => chore.nextDueDate != null).toList();
-
-  Future<void> completeChore(int index) async {
-    final Chore chore = homePageChores[index];
-
-    await adapter.complete(chore);
-
-    homePageChores.removeAt(index);
-
-    return refreshChores();
-  }
+      chores().where((chore) => chore.hasDueDate).toList();
 
   Future<void> refreshChores() async {
     isLoading(true);
 
     final List<Chore> newChores = await adapter.index();
-
     newChores.sort(_choresByDueDate);
-
     chores(newChores);
 
-    homePageChores(newChores.where((chore) {
-      return chore.hasNoDueDate && chore.nextDueUserId == auth.user()?.id;
-    }).toList());
+    _setFilterViews();
 
     isLoading(false);
   }
 
+  Future<void> completeChore(int choreId) async {
+    final int index = getIndexByChoreId(choreId);
+    final Chore chore = chores[index];
+
+    await adapter.complete(chore);
+
+    chores.removeAt(index);
+    _setFilterViews();
+
+    return refreshChores();
+  }
+
+  Future<void> snoozeChore(int choreId, DateTime date) async {
+    final int index = getIndexByChoreId(choreId);
+    final Chore chore = chores[index];
+
+    await adapter.snooze(chore, date);
+
+    chores.removeAt(index);
+    _setFilterViews();
+
+    return refreshChores();
+  }
+
+  void snoozeUntilTomorrow(int choreId) => snoozeChore(
+        choreId,
+        DateTimeExt.now().add(const Duration(days: 1)),
+      );
+
+  void snoozeUntilWeekend(int choreId) => snoozeChore(
+        choreId,
+        DateTimeExt.now().nextWeekend(),
+      );
+
+  int getIndexByChoreId(int choreId) =>
+      chores.indexWhere((chore) => chore.id == choreId);
+
+  void _setFilterViews() {
+    homePageChores(chores
+        .where((chore) => chore.hasDueDate && _choreBelongsToUser(chore))
+        .toList());
+  }
+
+  bool _choreBelongsToUser(Chore chore) =>
+      chore.nextDueUserId == auth.user()?.id;
+
   int _choresByDueDate(Chore choreA, Chore choreB) {
-    if (choreA.nextDueDate == null) {
-      if (choreB.nextDueDate == null) {
-        return 0;
-      }
-      return 1;
+    if (choreA.hasNoDueDate) {
+      return choreB.hasNoDueDate ? 0 : 1;
     }
 
-    if (choreB.nextDueDate == null) {
+    if (choreB.hasNoDueDate) {
       return -1;
     }
 
