@@ -1,20 +1,15 @@
-import 'dart:convert';
-
 import 'package:chore_manager_mobile/components/spinner.dart';
-import 'package:chore_manager_mobile/config/globals.dart';
 import 'package:chore_manager_mobile/constants/strings.dart';
 import 'package:chore_manager_mobile/modules/chores/chore.dart';
 import 'package:chore_manager_mobile/pages/home_page.dart';
+import 'package:chore_manager_mobile/pages/show_chore_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get/get.dart';
-import 'package:intl/intl.dart';
-import 'package:mocktail/mocktail.dart';
 
 import '../factories/chore_factory.dart';
 import '../helpers/helpers.dart';
 import '../mocks/data_mocks/chore_mocks.dart';
-import '../mocks/http_mocks.dart';
 import '../mocks/mocks.dart';
 
 void main() {
@@ -25,62 +20,15 @@ void main() {
 
     tearDown(Get.reset);
 
-    group('user sees chores with due date', () {
-      testWidgets('', (tester) async {
-        final chores = ChoreFactory().listOf(3);
-        mockChoreIndex(chores: chores);
-        await tester.pumpWidget(WidgetWrapper(HomePage()));
+    testWidgets('user sees chores with due date', (tester) async {
+      final chores = ChoreFactory().listOf(3);
+      mockChoreIndex(chores: chores);
+      await tester.pumpWidget(WidgetWrapper(HomePage()));
 
-        for (final chore in chores) {
-          expect(find.text(chore.title), findsOneWidget);
-        }
-      });
-
-      testWidgets('for today', (tester) async {
-        final chore = ChoreFactory().state({
-          'nextDueDate': DateTime.now(),
-        }).build();
-        mockChoreIndex(chores: [chore]);
-        await tester.pumpWidget(WidgetWrapper(HomePage()));
-
-        expect(find.text('today'), findsOneWidget);
-      });
-
-      testWidgets('for tomorrow', (tester) async {
-        final chore = ChoreFactory().state({
-          'nextDueDate': DateTime.now().add(const Duration(days: 1)),
-        }).build();
-        mockChoreIndex(chores: [chore]);
-        await tester.pumpWidget(WidgetWrapper(HomePage()));
-
-        expect(find.text('tomorrow'), findsOneWidget);
-      });
-
-      testWidgets('as day name for days between +2 and +6', (tester) async {
-        final dayPlus2 = DateTime.now().add(const Duration(days: 2));
-        final dayPlus6 = DateTime.now().add(const Duration(days: 6));
-        final chore1 = ChoreFactory().state({'nextDueDate': dayPlus2}).build();
-        final chore2 = ChoreFactory().state({'nextDueDate': dayPlus6}).build();
-
-        mockChoreIndex(chores: [chore1, chore2]);
-        await tester.pumpWidget(WidgetWrapper(HomePage()));
-
-        final dayPlus2String = DateFormat(DateFormat.WEEKDAY).format(dayPlus2);
-        final dayPlus6String = DateFormat(DateFormat.WEEKDAY).format(dayPlus6);
-        expect(find.text(dayPlus2String), findsOneWidget);
-        expect(find.text(dayPlus6String), findsOneWidget);
-      });
-
-      testWidgets('see date formatted for day 7 on', (tester) async {
-        final day = DateTime.now().add(const Duration(days: 7));
-        final chore = ChoreFactory().state({'nextDueDate': day}).build();
-
-        mockChoreIndex(chores: [chore]);
-        await tester.pumpWidget(WidgetWrapper(HomePage()));
-
-        final dayString = DateFormat(DateFormat.YEAR_NUM_MONTH_DAY).format(day);
-        expect(find.text(dayString), findsOneWidget);
-      });
+      for (final chore in chores) {
+        expect(find.text(chore.title), findsOneWidget);
+        expect(find.text(chore.friendlyDueDate), findsOneWidget);
+      }
     });
 
     testWidgets('user does not see chores with no due dates', (tester) async {
@@ -139,6 +87,25 @@ void main() {
       });
     });
 
+    group('tap on a chore', () {
+      late Chore chore;
+
+      setUp(() {
+        chore = ChoreFactory().build();
+        mockChoreIndex(chores: [chore]);
+      });
+
+      testWidgets('navigate to the show chore page', (tester) async {
+        await tester.pumpWidget(WidgetWrapper(HomePage()));
+
+        await tester.tap(find.text(chore.title));
+        await tester.pumpAndSettle();
+
+        expect(find.byType(ShowChorePage), findsOneWidget);
+        expect(find.text(chore.title), findsOneWidget);
+      });
+    });
+
     group('tap complete on a chore', () {
       late Chore chore;
 
@@ -170,11 +137,7 @@ void main() {
         await tester.dismiss(dismissible);
         await tester.pumpAndSettle();
 
-        verify(() => Globals.client.patch(
-              expectedPath('chores/${chore.id}'),
-              headers: expectedHeaders(),
-              body: jsonEncode({'completed': true}),
-            ));
+        verifyChoreComplete(chore: chore);
       });
 
       testWidgets('chore list is refreshed with new chores', (tester) async {
